Adding files via upload
Flask_Server_implementation works for sure with the sql databases.

Unity_Send_Func is a unity script that I believe you attach to one of the scenes. All that it needs to do is send the userID and the score they just got. The Flask_api code will handle whether or not to update.